+++ src/views/search-results-page/components/filters/components/checkbox.tsx
