+++ src/views/search-results-page/components/filters/components/date-slider/helpers.ts
