+++ src/views/search-results-page/components/filters/components/list.tsx
