+++ src/views/search-results-page/components/filters/components/section.tsx
